Mirroring GitHub into TFS, bringing (among other things) InProcess by default of GitHub release 0.5.2.
Testing pipeline [here](https://tfs.ansys.com:8443/tfs/ANSYS_Development/DPF/_build/results?buildId=7372183&view=logs&s=77b07857-e3de-563a-962e-fe3f8174b54b&j=32ca9294-3ac1-524a-049d-49e4f25e2670). OK

Related work items: #692152

+++ examples/00-basic/04-basic-load-file.py
@@ -52,19 +52,23 @@ export_csv_operator.run()
 # However, when the server file path is known, :func:`upload_file`
 # can be used.
 
-server_file_path = dpf.upload_file_in_tmp_folder(file_path)
-print(server_file_path)
+if not dpf.SERVER.local_server:
+    server_file_path = dpf.upload_file_in_tmp_folder(file_path)
+    print(server_file_path)
 
-# Remove file to avoid polluting.
-os.remove(file_path)
+    # Remove file to avoid polluting.
+    os.remove(file_path)
 
 ###############################################################################
 # Download CSV Result File
 # ~~~~~~~~~~~~~~~~~~~~~~~~~
 # Download the file ``simple_bar_fc.csv``:
 
-downloaded_client_file_path = os.getcwd() + "\\simple_bar_fc_downloaded.csv"
-dpf.download_file(server_file_path, downloaded_client_file_path)
+if not dpf.SERVER.local_server:
+    downloaded_client_file_path = os.getcwd() + "\\simple_bar_fc_downloaded.csv"
+    dpf.download_file(server_file_path, downloaded_client_file_path)
+else:
+    downloaded_client_file_path = file_path
 
 ###############################################################################
 # Load CSV Result File as Operator Input
